move brat libs to common
