+++ lib/commons/rscommons/reach_attributes.py
@@ -1,4 +1,4 @@
-from sqlbrat.lib.database import load_attributes
+from rscommons.database import load_attributes
 from rscommons.shapefile import write_attributes
 from osgeo import ogr
