+++ packages/brat/sqlbrat/brat_run.py
@@ -17,7 +17,8 @@ import time
 import datetime
 from osgeo import ogr
 from rscommons import Logger, initGDALOGRErrors, RSLayer, RSProject, ModelConfig, dotenv
-from sqlbrat.lib.database import execute_query, update_database, store_metadata
+from rscommons.database import execute_query, update_database, store_metadata, set_reach_fields_null
+from rscommons.reach_attributes import write_reach_attributes
 from sqlbrat.utils.vegetation_suitability import vegetation_suitability, output_vegetation_raster
 from sqlbrat.utils.vegetation_fis import vegetation_fis
 from sqlbrat.utils.combined_fis import combined_fis
@@ -25,8 +26,6 @@ from sqlbrat.utils.hydrology import hydrology
 from sqlbrat.utils.land_use import land_use
 from sqlbrat.utils.conservation import conservation
 from sqlbrat.utils.report import report
-from sqlbrat.lib.database import set_reach_fields_null
-from sqlbrat.lib.reach_attributes import write_reach_attributes
 from sqlbrat.__version__ import __version__
 
 cfg = ModelConfig('http://xml.riverscapes.xyz/Projects/XSD/V1/BRAT.xsd', __version__)
