+++ packages/brat/sqlbrat/utils/conservation.py
@@ -13,8 +13,8 @@ import argparse
 import sqlite3
 from osgeo import ogr
 from rscommons import Logger, dotenv
-from sqlbrat.lib.database import load_attributes
-from sqlbrat.lib.database import write_attributes
+from rscommons.database import load_attributes
+from rscommons.database import write_attributes
 import csv
